Models_Backend/Textfakedetection/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
groq>=0.4.0
python-dotenv==1.0.0
Pillow==10.1.0
numpy==1.24.3
easyocr==1.7.0
python-multipart==0.0.6
torch==2.0.1
torchvision==0.15.2
opencv-python-headless==4.8.1.78